audio_sentiment_analysis/requirements.txt
openai
deepgram-sdk
fastapi
uvicorn
python-dotenv